requirements.txt
streamlit==1.32.0
roboflow==1.1.14
supervision==0.18.0
opencv-python==4.9.0.80
numpy==1.26.4
Pillow==10.2.0 
python-dotenv==1.1.0
